Let the jump timer be signed on both views of a car
The player car and a rival share their layout field for field: both are
412 bytes, sixty field names in common, and every one of them at the
same offset and the same size, which is what makes AsRivalCar safe.
One field disagreed on its type rather than its place. The jump timer
was u16 on the player and s16 on the rival, so the two views read the
same sixteen bits and meant different things by them, and neither the
compiler nor the union conversion can see that.

The code was already paying for it. Every use on the player side cast
back to s16 to get the value it wanted, and its three neighbours in the
same group, the state, the rate and the target height, are all s16. The
player declaration was simply the odd one out.

The two casts that became identities are gone. The cast in the jump arc
stays: it truncates an s32 and is what makes the counter wrap at
thirty-two thousand rather than run away.

+++ src/main/PAL/main/car/update_player_car.c
@@ -370,10 +370,10 @@ static void LandFromJump(PlayerCarRuntime *car, GameCarDrive *p, s32 ground) {
     car->verticalRoll = 0;
     StartCarBodyKick(1, AsRivalCar(car));
     g_ShiftSoundLevel = 0;
-    if (((s16)car->verticalMotionTimer >= 19) && (g_RacePhase < 3)) {
+    if ((car->verticalMotionTimer >= 19) && (g_RacePhase < 3)) {
         PlaySoundCue(0xE);
     }
-    if ((p->motionState == CAR_MOTION_DRIVING) && ((s16)car->verticalMotionTimer >= 3)) {
+    if ((p->motionState == CAR_MOTION_DRIVING) && (car->verticalMotionTimer >= 3)) {
         RelaunchDrivetrain(car, p);
     }
 }
